demangle: add LambdaAuto for lambda auto parameters
Don't try to reuse TemplateParam and distinguish at print time.
Instead, distinguish entirely when converting to AST.

To make this work, convert back and forth between TemplateParam and
LambdaAuto when copying a substituted value.  Keep track of Template
values as we copy.

Add a test case that used to crash in ASTToString.