Make sure to not modify the input options slice.
In Go, appending to a slice can modify the original slice if there is
enough backing space. In use cases like passing options slice to an API
call this is unexpected since it's quite intuitive at the API surface to
write a loop to demangle multiple names and reuse the same options slice
between the calls. Forcing clients to copy or re-create the options in
such pattern seems unnecessary.